Manager: fix comma-separated display of >32 bit numbers
VS apparently requires "long long" to get 64 bits.

Also: convert various floats to double in the Manager.
BOINC shouldn't use floats anywhere.  Always use doubles.

+++ clientgui/MainDocument.cpp
@@ -2629,8 +2629,8 @@ void color_cycle(int i, int n, wxColour& color) {
 }
 
 #ifdef __WXMSW__
-static float XDPIScaleFactor = 0.0;
-static float YDPIScaleFactor = 0.0;
+static double XDPIScaleFactor = 0.0;
+static double YDPIScaleFactor = 0.0;
 
 void GetDPIScaling() {
 	XDPIScaleFactor = 1.0;
@@ -2650,14 +2650,14 @@ void GetDPIScaling() {
 	FreeLibrary(hUser32);
 }
 
-float GetXDPIScaling() {
+double GetXDPIScaling() {
 	if (XDPIScaleFactor == 0.0) {
 		GetDPIScaling();
 	}
 	return XDPIScaleFactor;
 }
 
-float GetYDPIScaling() {
+double GetYDPIScaling() {
 	if (YDPIScaleFactor == 0.0) {
 		GetDPIScaling();
 	}
